fix FFI
