+++ delta-kernel-engine-ffi/Cargo.toml
@@ -11,7 +11,7 @@ version.workspace = true
 
 [dependencies]
 delta_kernel = { path = "../kernel" }
-delta_kernel_ffi = { path = "../ffi" }
+delta_kernel_ffi = { path = "../ffi", features = ["internal-api"] }
 delta_kernel_ffi_macros = { path = "../ffi-proc-macros" }
 delta-kernel-engine = { path = "../delta-kernel-engine" }
 tracing = "0.1"
