+++ ffi/src/lib.rs
@@ -6,13 +6,14 @@ use std::default::Default;
 use std::os::raw::{c_char, c_void};
 use std::ptr::NonNull;
 use std::sync::Arc;
-use tracing::debug;
-use url::Url;
 
 use delta_kernel::schema::Schema;
 use delta_kernel::snapshot::Snapshot;
 use delta_kernel::{DeltaResult, Engine, EngineData, Table};
 use delta_kernel_ffi_macros::handle_descriptor;
+
+use tracing::debug;
+use url::Url;
 
 // cbindgen doesn't understand our use of feature flags here, and by default it parses `mod handle`
 // twice. So we tell it to ignore one of the declarations to avoid double-definition errors.
@@ -144,7 +145,7 @@ macro_rules! kernel_string_slice {
 }
 pub(crate) use kernel_string_slice;
 
-trait TryFromStringSlice<'a>: Sized {
+pub trait TryFromStringSlice<'a>: Sized {
     unsafe fn try_from_slice(slice: &'a KernelStringSlice) -> DeltaResult<Self>;
 }
 
@@ -353,14 +354,12 @@ pub trait ExternEngine: Send + Sync {
 #[handle_descriptor(target=dyn ExternEngine, mutable=false)]
 pub struct SharedExternEngine;
 
-#[cfg(any(feature = "default-engine", feature = "sync-engine"))]
 struct ExternEngineVtable {
     // Actual engine instance to use
     engine: Arc<dyn Engine>,
     allocate_error: AllocateErrorFn,
 }
 
-#[cfg(any(feature = "default-engine", feature = "sync-engine"))]
 impl Drop for ExternEngineVtable {
     fn drop(&mut self) {
         debug!("dropping engine interface");
@@ -371,7 +370,6 @@ impl Drop for ExternEngineVtable {
 ///
 /// Kernel doesn't use any threading or concurrency. If engine chooses to do so, engine is
 /// responsible for handling  any races that could result.
-#[cfg(any(feature = "default-engine", feature = "sync-engine"))]
 unsafe impl Send for ExternEngineVtable {}
 
 /// # Safety
@@ -383,10 +381,8 @@ unsafe impl Send for ExternEngineVtable {}
 /// Basically, by failing to implement these traits, we forbid the engine from being able to declare
 /// its thread-safety (because rust assumes it is not threadsafe). By implementing them, we leave it
 /// up to the engine to enforce thread safety if engine chooses to use threads at all.
-#[cfg(any(feature = "default-engine", feature = "sync-engine"))]
 unsafe impl Sync for ExternEngineVtable {}
 
-#[cfg(any(feature = "default-engine", feature = "sync-engine"))]
 impl ExternEngine for ExternEngineVtable {
     fn engine(&self) -> Arc<dyn Engine> {
         self.engine.clone()
@@ -399,25 +395,13 @@ impl ExternEngine for ExternEngineVtable {
 /// # Safety
 ///
 /// Caller is responsible for passing a valid path pointer.
-unsafe fn unwrap_and_parse_path_as_url(path: KernelStringSlice) -> DeltaResult<Url> {
+pub unsafe fn unwrap_and_parse_path_as_url(path: KernelStringSlice) -> DeltaResult<Url> {
     let path: &str = unsafe { TryFromStringSlice::try_from_slice(&path) }?;
     let table = Table::try_from_uri(path)?;
     Ok(table.location().clone())
 }
 
-/// # Safety
-///
-/// Caller is responsible for passing a valid path pointer.
-#[cfg(feature = "sync-engine")]
-#[no_mangle]
-pub unsafe extern "C" fn get_sync_engine(
-    allocate_error: AllocateErrorFn,
-) -> ExternResult<Handle<SharedExternEngine>> {
-    get_sync_engine_impl(allocate_error).into_extern_result(&allocate_error)
-}
-
-#[cfg(any(feature = "default-engine", feature = "sync-engine"))]
-fn engine_to_handle(
+pub fn engine_to_handle(
     engine: Arc<dyn Engine>,
     allocate_error: AllocateErrorFn,
 ) -> Handle<SharedExternEngine> {
@@ -426,30 +410,6 @@ fn engine_to_handle(
         allocate_error,
     });
     engine.into()
-}
-
-#[cfg(feature = "default-engine")]
-fn get_default_engine_impl(
-    url: Url,
-    options: HashMap<String, String>,
-    allocate_error: AllocateErrorFn,
-) -> DeltaResult<Handle<SharedExternEngine>> {
-    use delta_kernel::engine::default::executor::tokio::TokioBackgroundExecutor;
-    use delta_kernel::engine::default::DefaultEngine;
-    let engine = DefaultEngine::<TokioBackgroundExecutor>::try_new(
-        &url,
-        options,
-        Arc::new(TokioBackgroundExecutor::new()),
-    );
-    Ok(engine_to_handle(Arc::new(engine?), allocate_error))
-}
-
-#[cfg(feature = "sync-engine")]
-fn get_sync_engine_impl(
-    allocate_error: AllocateErrorFn,
-) -> DeltaResult<Handle<SharedExternEngine>> {
-    let engine = delta_kernel::engine::sync::SyncEngine::new();
-    Ok(engine_to_handle(Arc::new(engine), allocate_error))
 }
 
 /// # Safety
@@ -666,8 +626,8 @@ impl<T> Default for ReferenceSet<T> {
 
 #[cfg(test)]
 mod tests {
-    use delta_kernel::engine::default::{executor::tokio::TokioBackgroundExecutor, DefaultEngine};
-    use delta_kernel::object_store::memory::InMemory;
+    use delta_kernel_engine::default::{executor::tokio::TokioBackgroundExecutor, DefaultEngine};
+    use delta_kernel_engine::object_store::memory::InMemory;
     use test_utils::{actions_to_string, actions_to_string_partitioned, add_commit, TestAction};
 
     use super::*;
@@ -795,15 +755,5 @@ mod tests {
         unsafe { free_snapshot(snapshot) }
         unsafe { free_engine(engine) }
         Ok(())
-    }
-
-    #[test]
-    #[cfg(feature = "sync-engine")]
-    fn sync_engine() {
-        let engine = unsafe { get_sync_engine(allocate_err) };
-        let engine = ok_or_panic(engine);
-        unsafe {
-            free_engine(engine);
-        }
     }
 }
